Refactor datasets package (empty __init__.py)
