+++ lib/datasets/factory.py
@@ -9,14 +9,14 @@
 
 __sets = {}
 
-import datasets.pascal_voc
+from datasets.pascal_voc import pascal_voc
 import numpy as np
 
 def _selective_search_IJCV_top_k(split, year, top_k):
     """Return an imdb that uses the top k proposals from the selective search
     IJCV code.
     """
-    imdb = datasets.pascal_voc(split, year)
+    imdb = pascal_voc(split, year)
     imdb.roidb_handler = imdb.selective_search_IJCV_roidb
     imdb.config['top_k'] = top_k
     return imdb
@@ -26,7 +26,7 @@ for year in ['2007', '2012']:
     for split in ['train', 'val', 'trainval', 'test']:
         name = 'voc_{}_{}'.format(year, split)
         __sets[name] = (lambda split=split, year=year:
-                datasets.pascal_voc(split, year))
+                pascal_voc(split, year))
 
 # Set up voc_<year>_<split>_top_<k> using selective search "quality" mode
 # but only returning the first k boxes
