+++ tools/train_faster_rcnn_alt_opt.py
@@ -18,7 +18,6 @@ from fast_rcnn.train import get_training_roidb, train_net
 from fast_rcnn.config import cfg, cfg_from_file, cfg_from_list, get_output_dir
 from datasets.factory import get_imdb
 from rpn.generate import imdb_proposals
-import datasets.imdb
 import argparse
 import pprint
 import numpy as np
